+++ src/templates/MyProject.Angular/Program.cs
@@ -31,4 +31,4 @@ app.UseSimplifyWebNonTerminal();
 if (!app.Environment.IsDevelopment())
 	app.UseSpa(spa => spa.Options.SourcePath = "ClientApp");
 
-app.Run();
+await app.RunAsync();