+++ src/templates/MyProject.Api/Setup/IocRegistrations.cs
@@ -1,7 +1,6 @@
 ﻿using MyProject.Api.ViewModels;
 using Simplify.DI;
 using Simplify.Web;
-using Simplify.Web.Json;
 
 namespace MyProject.Api.Setup;
 
@@ -12,7 +11,6 @@ public static class IocRegistrations
 		// Simplify.DI.DIContainer.Current IOC container registrations starting point
 
 		provider.RegisterSimplifyWeb()
-			.RegisterJsonModelBinder()
 
 		.Register<SampleModelFactory>();