+++ src/templates/MyProject.Api/Startup.cs
@@ -1,8 +1,6 @@
 ﻿using MyProject.Api.Setup;
 using Simplify.DI;
 using Simplify.Web;
-using Simplify.Web.Json.Model.Binding;
-using Simplify.Web.Model;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -15,9 +13,6 @@ var app = builder.Build();
 if (app.Environment.IsDevelopment())
 	app.UseDeveloperExceptionPage();
 
-// Enabling Simplify.Web JSON requests handling
-HttpModelHandler.RegisterModelBinder<JsonModelBinder>();
+app.UseSimplifyWeb();
 
-app.UseSimplifyWebWithoutRegistrations();
-
-app.Run();
+await app.RunAsync();