+++ src/templates/MyProject.SchedulerApi/Setup/IocRegistrations.cs
@@ -1,7 +1,6 @@
 ﻿using MyProject.SchedulerApi.Settings;
 using Simplify.DI;
 using Simplify.Web;
-using Simplify.Web.Json;
 
 namespace MyProject.SchedulerApi.Setup;
 
@@ -10,7 +9,6 @@ public static class IocRegistrations
 	public static IDIContainerProvider RegisterAll(this IDIContainerProvider provider)
 	{
 		provider.RegisterSimplifyWeb()
-			.RegisterJsonModelBinder()
 
 		.Register(r => new WebApplicationStartupSettings(r.Resolve<IConfiguration>()), LifetimeType.Singleton)
 		.Register<WebApplicationStartup>(LifetimeType.Singleton);