+++ src/templates/MyProject.SchedulerApi/WebApplicationStartup.cs
@@ -1,31 +1,24 @@
 ﻿using MyProject.SchedulerApi.Settings;
 using Simplify.Scheduler.Jobs;
 using Simplify.Web;
-using Simplify.Web.Json.Model.Binding;
-using Simplify.Web.Model;
 
 namespace MyProject.SchedulerApi;
 
 public class WebApplicationStartup(WebApplicationStartupSettings settings)
 {
-	private readonly WebApplicationStartupSettings _settings = settings;
-
-	public void Run(IJobArgs args)
+	public async Task Run(IJobArgs args)
 	{
 		var builder = WebApplication.CreateBuilder((string[])args.StartupArgs!);
 
-		builder.WebHost.UseUrls($"http://{_settings.BindHostName}:{_settings.WorkingPort}");
+		builder.WebHost.UseUrls($"http://{settings.BindHostName}:{settings.WorkingPort}");
 
 		var app = builder.Build();
 
 		if (app.Environment.IsDevelopment())
 			app.UseDeveloperExceptionPage();
 
-		// Enabling Simplify.Web JSON requests handling
-		HttpModelHandler.RegisterModelBinder<JsonModelBinder>();
+		app.UseSimplifyWeb();
 
-		app.UseSimplifyWebWithoutRegistrations();
-
-		app.Run();
+		await app.RunAsync();
 	}
 }