+++ src/templates/MyProject.WindowsServiceApi/Setup/IocRegistrations.cs
@@ -2,7 +2,6 @@
 using MyProject.WindowsServiceApi.Settings;
 using Simplify.DI;
 using Simplify.Web;
-using Simplify.Web.Json;
 
 namespace MyProject.WindowsServiceApi.Setup;
 
@@ -11,7 +10,6 @@ public static class IocRegistrations
 	public static IDIContainerProvider RegisterAll(this IDIContainerProvider provider)
 	{
 		provider.RegisterSimplifyWeb()
-			.RegisterJsonModelBinder()
 
 		.Register(r => new WebApplicationStartupSettings(r.Resolve<IConfiguration>()), LifetimeType.Singleton)
 		.Register<WebApplicationStartup>(LifetimeType.Singleton);